requirements.txt
# Deep Learning
tensorflow>=2.13.0
keras>=2.13.0

# Computer Vision
opencv-python>=4.8.0
opencv-contrib-python>=4.8.0

# Data Science
numpy>=1.24.0
pandas>=2.0.0
matplotlib>=3.7.0
seaborn>=0.12.0
scikit-learn>=1.3.0

# Data Augmentation
albumentations>=1.3.0

# Visualization
plotly>=5.14.0

# Progress bars
tqdm>=4.65.0

# Configuration
pyyaml>=6.0
python-dotenv>=1.0.0

# Jupyter
jupyter>=1.0.0
ipykernel>=6.25.0

# Future: Web Interface
# streamlit>=1.25.0
# gradio>=3.40.0

# Development
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
